requirements.txt
asgiref==3.4.1
colorama==0.4.6
distlib==0.3.6
Django==4.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-fluent-utils==3.0.1
django-parler==2.3
django-polymorphic==3.1.0
django-tag-parser==3.2
django-template-analyzer==2.0
filelock==3.8.0
html5lib==1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.0
Pillow==9.3.0
platformdirs==2.5.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
Unidecode==1.3.6
virtualenv==20.16.6
webencodings==0.5.1
